Refactor so each system class is in its own file
