+++ src/main/kotlin/com/mineinabyss/looty/features/recipes/ItemRecipeQuery.kt
@@ -1,7 +1,5 @@
 package com.mineinabyss.looty.features.recipes
 
-import com.mineinabyss.geary.papermc.datastore.decodePrefabs
-import com.mineinabyss.geary.papermc.tracking.items.components.SetItem
 import com.mineinabyss.geary.papermc.tracking.items.gearyItems
 import com.mineinabyss.geary.prefabs.PrefabKey
 import com.mineinabyss.geary.systems.accessors.TargetScope
@@ -10,11 +8,6 @@ import com.mineinabyss.idofront.recipes.register
 import com.mineinabyss.looty.config.looty
 import org.bukkit.Bukkit
 import org.bukkit.NamespacedKey
-import org.bukkit.event.EventHandler
-import org.bukkit.event.EventPriority
-import org.bukkit.event.Listener
-import org.bukkit.event.inventory.PrepareItemCraftEvent
-import org.bukkit.event.player.PlayerJoinEvent
 import org.bukkit.inventory.ItemStack
 
 class ItemRecipeQuery : Query() {
@@ -45,25 +38,5 @@ class ItemRecipeQuery : Query() {
             }.onFailure { it.printStackTrace() }
         }
         return discoveredRecipes
-    }
-}
-
-class RecipeDiscoverySystem(val discoveredRecipes: List<NamespacedKey>): Listener {
-    @EventHandler
-    fun PlayerJoinEvent.showRecipesOnJoin() {
-        player.discoverRecipes(discoveredRecipes)
-    }
-}
-
-class RecipeCraftingSystem : Listener {
-    /**
-     * Prevents custom items being usable in vanilla recipes based on their material,
-     * when they have a [DenyInVanillaRecipes] component, by setting result to null.
-     */
-    @EventHandler(priority = EventPriority.HIGHEST)
-    fun PrepareItemCraftEvent.onCraftWithCustomItem() {
-        if (inventory.matrix.asSequence().mapNotNull {
-                it?.itemMeta?.persistentDataContainer?.decodePrefabs()?.firstOrNull()?.toEntityOrNull()
-            }.any { it.has<SetItem>() && it.has<DenyInVanillaRecipes>() }) inventory.result = null
     }
 }
